+++ R/artificial_clustering.R
@@ -5,7 +5,7 @@ artificial_clustering <- function(points) {
  l_infty <- max(abs(points))
 
  ret <- list()
- ret$height <- rep(l_infty, length(points)-1) + (1e-10)*seq_along(points[-1])
+ ret$height <- rep(l_infty, length(points)-1) + l_infty*(1e-10)*seq_along(points[-1])
  ret$order_points <- seq_along(points)
  ret$labels <- names(points)
  ret$method <- "artificial"